Split Main and Navbar
